refactor: extract core Angular module
